pyproject.toml
[project]
name = "jira-to-code-env"
version = "0.1.0"
description = "An environment where agents resolve Jira tickets by reading, writing, and testing code."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn",
    "pytest",
    "openai",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_to_code"]

[project.scripts]
server = "server.app:main"